+++ ui/src/main/java/com/jin/ui/cart/content/CartHeader.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.cart.ui.content
+package com.jin.ui.cart.content
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
@@ -13,11 +13,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jin.honey.R
-import com.jin.ui.theme.HoneyTheme
+import com.jin.ui.R
 
 @Composable
 fun CartHeader() {
